Extract shared utilities
